Prepare Release 3.0.2-RC2
