+++ datasketches-memory-java8/src/main/java/org/apache/datasketches/memory/internal/UnsafeUtil.java
@@ -32,8 +32,6 @@ import sun.misc.Unsafe;
 @SuppressWarnings({"restriction","javadoc"})
 public final class UnsafeUtil {
   public static final Unsafe unsafe;
-  public static final String JDK; //must be at least "1.8"
-  public static final int JDK_MAJOR; //8, 9, 10, 11, 12, etc
 
   //not an indicator of whether compressed references are used.
   public static final int ADDRESS_SIZE;
@@ -115,50 +113,9 @@ public final class UnsafeUtil {
 
     ARRAY_OBJECT_INDEX_SCALE = unsafe.arrayIndexScale(Object[].class);
     OBJECT_SHIFT = ARRAY_OBJECT_INDEX_SCALE == 4 ? 2 : 3;
-
-    final String jdkVer = System.getProperty("java.version");
-    final int[] p = parseJavaVersion(jdkVer);
-    JDK = p[0] + "." + p[1];
-    JDK_MAJOR = (p[0] == 1) ? p[1] : p[0];
   }
 
   private UnsafeUtil() {}
-
-  /**
-   * Returns first two number groups of the java version string.
-   * @param jdkVer the java version string from System.getProperty("java.version").
-   * @return first two number groups of the java version string.
-   */
-  public static int[] parseJavaVersion(final String jdkVer) {
-    final int p0, p1;
-    try {
-      String[] parts = jdkVer.trim().split("[^0-9\\.]");//grab only number groups and "."
-      parts = parts[0].split("\\."); //split out the number groups
-      p0 = Integer.parseInt(parts[0]); //the first number group
-      p1 = (parts.length > 1) ? Integer.parseInt(parts[1]) : 0; //2nd number group, or 0
-    } catch (final NumberFormatException | ArrayIndexOutOfBoundsException  e) {
-      throw new IllegalArgumentException("Improper Java -version string: " + jdkVer + LS + e);
-    }
-    checkJavaVersion(jdkVer, p0, p1);
-    return new int[] {p0, p1};
-  }
-
-  /**
-   * Checks the runtime Java Version string. Note that Java 17 and 21 is allowed only because some clients do not use the
-   * WritableMemory.allocateDirect(..) and related functions, which will not work with Java versions >= 14.  
-   * The on-heap functions may work with 17 and 21, nonetheless, versions > Java 11 are not officially supported. 
-   * Caveat emptor.
-   * @param jdkVer the <i>System.getProperty("java.version")</i> string of the form "p0.p1.X"
-   * @param p0 The first number group 
-   * @param p1 The second number group
-   */
-  public static void checkJavaVersion(final String jdkVer, final int p0, final int p1) {
-    final boolean ok = ( ((p0 == 1) && (p1 == 8)) || (p0 == 8) || (p0 == 11) || (p0 == 17) || (p0 == 21));
-    if (!ok) {
-      throw new IllegalArgumentException(
-          "Unsupported Runtime JDK Major Version, must be one of 1.8, 8, 11, 17, 21: " + jdkVer);
-    }
-  }
 
   /**
    * Gets the <i>unsafe.objectFieldOffset(..)</i> for declared fields of a class.
