+++ datasketches-memory-java8/src/test/java/org/apache/datasketches/memory/internal/UnsafeUtilTest.java
@@ -20,7 +20,6 @@
 package org.apache.datasketches.memory.internal;
 
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
 import java.util.ArrayList;
@@ -33,63 +32,6 @@ import org.testng.annotations.Test;
  */
 public class UnsafeUtilTest {
   long testField = 1; //Do not remove & cannot be static. Used in reflection check.
-
-  @Test
-  public void checkJdkString() {
-    String jdkVer;
-    int[] p = new int[2];
-    String[] good1_Strings = {"1.8.0_121", "8", "11", "17"};
-    int len = good1_Strings.length;
-    for (int i = 0; i < len; i++) {
-      jdkVer = good1_Strings[i];
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]);
-      int jdkMajor = (p[0] == 1) ? p[1] : p[0]; //model the actual JDK_MAJOR
-      if (p[0] == 1) { assertTrue(jdkMajor == p[1]); }
-      if (p[0] > 1 ) { assertTrue(jdkMajor == p[0]); }
-    }
-    try {
-      jdkVer = "14.0.4"; //ver 14 string
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]);
-      fail();
-    } catch (IllegalArgumentException e) {
-      println("" + e);
-    }
-
-    try {
-      jdkVer = "1.7.0_80"; //1.7 string
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]);
-      fail();
-    } catch (IllegalArgumentException e) {
-      println("" + e);
-    }
-    try {
-      jdkVer = "1.6.0_65"; //valid string but < 1.7
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]); //throws
-      fail();
-    } catch (IllegalArgumentException e) {
-      println("" + e);
-    }
-    try {
-      jdkVer = "b"; //invalid string
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]); //throws
-      fail();
-    } catch (IllegalArgumentException e) {
-      println("" + e);
-    }
-    try {
-      jdkVer = ""; //invalid string
-      p = UnsafeUtil.parseJavaVersion(jdkVer);
-      UnsafeUtil.checkJavaVersion(jdkVer, p[0], p[1]); //throws
-      fail();
-    } catch (IllegalArgumentException e) {
-      println("" + e);
-    }
-  }
 
   @Test
   public void checkFieldOffset() {
